refactor(graphql-client): extract shared scalars and fragments into common module
Move impl_scalar! declarations and the cross-cutting cynic fragments
(Base64, BigInt, DateTime, MoveData, GQLAddress, MoveObject,
MoveObjectContents, MoveValue, MoveType, PageInfo) from query_types/mod.rs
into a dedicated common.rs. No public API changes — every type is
re-exported from query_types so existing imports continue to resolve.

Part of #512.
