+++ crates/iota-sdk-graphql-client/src/query_types/mod.rs
@@ -1,5 +1,5 @@
 // Copyright (c) Mysten Labs, Inc.
-// Modifications Copyright (c) 2025 IOTA Stiftung
+// Modifications Copyright (c) 2026 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
 mod active_validators;
@@ -7,6 +7,7 @@ mod balance;
 mod chain;
 mod checkpoint;
 mod coin;
+mod common;
 mod dry_run;
 mod dynamic_fields;
 mod epoch;
@@ -32,7 +33,10 @@ pub use checkpoint::{
     CheckpointsQuery,
 };
 pub use coin::{CoinMetadata, CoinMetadataArgs, CoinMetadataQuery};
-use cynic::impl_scalar;
+pub use common::{
+    Address, Base64, BigInt, DateTime, GQLAddress, JsonValue, MoveData, MoveObject,
+    MoveObjectContents, MoveType, MoveValue, ObjectId, PageInfo,
+};
 pub use dry_run::{
     DryRunArgs, DryRunEffect, DryRunMutation, DryRunQuery, DryRunResult, DryRunReturn, GasCoin,
     Input, ObjectRef, ResultArg, TransactionArgument, TransactionMetadata,
@@ -50,7 +54,6 @@ pub use iota_names::{
     NameRegistration, NameRegistrationConnection, ResolveIotaNamesAddressArgs,
     ResolveIotaNamesAddressQuery,
 };
-use iota_types::{Address, ObjectId};
 pub use move_view_call::{MoveViewCallArgs, MoveViewCallQuery, MoveViewResult};
 pub use normalized_move::{
     MoveAbility, MoveEnum, MoveEnumConnection, MoveEnumVariant, MoveField, MoveFunction,
@@ -71,7 +74,6 @@ pub use protocol_config::{
     ProtocolConfigAttr, ProtocolConfigFeatureFlag, ProtocolConfigQuery, ProtocolConfigs,
     ProtocolVersionArgs,
 };
-use serde_json::Value as JsonValue;
 pub use service_config::{Feature, ServiceConfig, ServiceConfigQuery};
 pub use transaction::{
     TransactionBlock, TransactionBlockArgs, TransactionBlockCheckpointQuery,
@@ -81,94 +83,5 @@ pub use transaction::{
     TransactionBlocksWithEffectsQuery, TransactionsFilter,
 };
 
-use crate::error;
-
 #[cynic::schema("rpc")]
 pub mod schema {}
-
-// ===========================================================================
-// Scalars
-// ===========================================================================
-
-impl_scalar!(Address, schema::IotaAddress);
-impl_scalar!(ObjectId, schema::IotaAddress);
-impl_scalar!(u64, schema::UInt53);
-impl_scalar!(JsonValue, schema::JSON);
-
-#[derive(Clone, cynic::Scalar, Debug, derive_more::From)]
-#[cynic(graphql_type = "Base64")]
-pub struct Base64(pub String);
-
-#[derive(Clone, cynic::Scalar, Debug, derive_more::From)]
-#[cynic(graphql_type = "BigInt")]
-pub struct BigInt(pub String);
-
-#[derive(Clone, cynic::Scalar, Debug)]
-#[cynic(graphql_type = "DateTime")]
-pub struct DateTime(pub String);
-
-#[derive(Clone, cynic::Scalar, Debug, derive_more::From)]
-#[cynic(graphql_type = "MoveData")]
-pub struct MoveData(pub serde_json::Value);
-
-// ===========================================================================
-// Types used in several queries
-// ===========================================================================
-
-#[derive(Clone, Copy, cynic::QueryFragment, Debug)]
-#[cynic(schema = "rpc", graphql_type = "Address")]
-pub struct GQLAddress {
-    pub address: Address,
-}
-
-#[derive(Clone, cynic::QueryFragment, Debug)]
-#[cynic(schema = "rpc", graphql_type = "MoveObject")]
-pub struct MoveObject {
-    pub bcs: Option<Base64>,
-}
-
-#[derive(cynic::QueryFragment, Debug)]
-#[cynic(schema = "rpc", graphql_type = "MoveObject")]
-pub struct MoveObjectContents {
-    pub contents: Option<MoveValue>,
-}
-
-#[derive(cynic::QueryFragment, Debug)]
-#[cynic(schema = "rpc", graphql_type = "MoveValue")]
-pub struct MoveValue {
-    pub type_: MoveType,
-    pub bcs: Base64,
-    pub json: Option<JsonValue>,
-}
-
-#[derive(Clone, cynic::QueryFragment, Debug)]
-#[cynic(schema = "rpc", graphql_type = "MoveType")]
-pub struct MoveType {
-    pub repr: String,
-}
-
-// ===========================================================================
-// Utility Types
-// ===========================================================================
-
-#[derive(Clone, cynic::QueryFragment, Debug, Default)]
-#[cynic(schema = "rpc", graphql_type = "PageInfo")]
-/// Information about pagination in a connection.
-pub struct PageInfo {
-    /// When paginating backwards, are there more items?
-    pub has_previous_page: bool,
-    /// Are there more items when paginating forwards?
-    pub has_next_page: bool,
-    /// When paginating backwards, the cursor to continue.
-    pub start_cursor: Option<String>,
-    /// When paginating forwards, the cursor to continue.
-    pub end_cursor: Option<String>,
-}
-
-impl TryFrom<BigInt> for u64 {
-    type Error = error::Error;
-
-    fn try_from(value: BigInt) -> Result<Self, Self::Error> {
-        Ok(value.0.parse::<u64>()?)
-    }
-}
